+++ libs/components/src/lib/modules/overlay/overlay-control.ts
@@ -7,18 +7,16 @@ import {
   Injector,
   ViewRef,
 } from '@angular/core';
-import { animationFrameScheduler, fromEvent, merge, merge as mergeObs, Observable, Subject } from 'rxjs';
+import { animationFrameScheduler, fromEvent, merge as mergeObs, Observable, Subject } from 'rxjs';
 import {
   debounceTime,
   distinctUntilChanged,
   filter,
   map,
-  mapTo,
   observeOn,
   skipWhile,
   takeUntil,
   tap,
-  throttleTime,
 } from 'rxjs/operators';
 import {
   PrizmOverlayConfig,
@@ -32,6 +30,7 @@ import { PrizmOverlayAbstractPosition } from './position/position';
 import { PrizmOverlayComponent } from './overlay.component';
 import { BODY_ELEMENT, EventBus, getContent } from './utils';
 import { WINDOW } from '@ng-web-apis/common';
+import { raceEmit } from '@prizm-ui/helpers';
 
 export class PrizmOverlayControl {
   position: PrizmOverlayAbstractPosition;
@@ -105,13 +104,12 @@ export class PrizmOverlayControl {
     const insideClick = fromEvent<MouseEvent>(this.viewEl.querySelector('.z-overlay-wrapper'), 'click');
     const outsideClick = fromEvent<MouseEvent>(this.viewEl, 'click');
 
-    // For detect dynamic html elements
-    return merge<[[MouseEvent, boolean], [MouseEvent, boolean]]>(
+    return raceEmit<[MouseEvent, boolean]>(
+      100,
       insideClick.pipe(map((e: MouseEvent) => [e, false])),
       outsideClick.pipe(map((e: MouseEvent) => [e, true]))
     ).pipe(
-      throttleTime(100),
-      filter(([, isOtsideClick]) => isOtsideClick),
+      filter(([, isOutsideClick]) => isOutsideClick),
       map(([e]: [MouseEvent, boolean]) => e.target),
       skipWhile(() => !this.config.closeOnDocClick),
       tap(() => {
@@ -119,18 +117,6 @@ export class PrizmOverlayControl {
         this.close();
       })
     );
-
-    // return fromEvent(this.viewEl, 'click').pipe(
-    //   takeUntil(this.destroy$),
-    //   map((e: any) => e.target),
-    //   skipWhile(() => !this.config.closeOnDocClick),
-    //   filter(this.isNotHostElement.bind(this)),
-    //   tap(() => {
-    //     console.log('#mz onDocumentClick');
-    //     this.config.docClickCallback();
-    //     this.close();
-    //   })
-    // );
   }
 
   public onWindowResize(): Observable<any> {