+++ apps/doc/src/app/about-prizm/changelog/CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Features
 
 - feat(components/overlay): added new logic for detecting events from dynamic elements #411
+- feat(helpers/rxjs): 'raceEmit' operator was added to detect first emit in the transmitted time interval between source streams
 
 ### Bug fixes